ARM: make kernel oops easier to read
We don't need the offset for the first function name in each backtrace
entry; this needlessly consumes screen space.  This is virtually always
the first or second instruction in the called function.

Also, recognise stmfd instructions which include r10 as a valid stack
saving instruction, and when dumping the registers, dump six registers
per line rather than five, and fix the wrapping.

Change-Id: Ifc58cf95a3458bed86e67299a2f7a843de6d1618
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>