power: qpnp-bms: qpnp-charger: make logs easier to parse
There are usescases where the BMS and charger kernel logs are being
parsed for real time driver operation updates in performance testing.
There are a few suggested changes to the logging in order to make
getting the relevant information easier.

The precompensated battery current values are very useful for the VI
team to verify the battery current temperature compensation scheme.
Print it along with the final battery current.

Also, add a comma between the bat temperature logs in order to better
distinguish between the adc code value and the degC value.

Change-Id: Ie66eb7ed6e030cf15f0e6ebe3fd23c39deb9ec51
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>