selinux: add a ugly workaround to bail early
when selinux tries to free a socket that is still in use and it
ends in a Kernel panic.

These lines are from opensource.samsung.com.
While this will most likely put a bandaid over this Kernel panic it does not
fix the bug so it might occur somewhere else if the same conditions are met
but since I have only seen panics from this hook then we may as well sleep in peace.

Change-Id: I68f0842232d6527cfd7dce48c640911154a621c0
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>