sched: Fix inaccurate accounting for real-time task
It is possible that rq->clock_task was not updated in put_prev_task()
in which case we can potentially overcharge a real-time task for time
it did not run. This is because clock_task could be stale and not
represent the exact time real-time task started running.

Fix this by forcing update of rq->clock_task when real-time task
starts running.

Change-Id: Ie7af0702cca847ca7ca192c5c12fe38f8436d7cc
Signed-off-by: Srivatsa Vaddagiri <vatsa@codeaurora.org>