video: mdss_mdp_pp: Check mutex state prior blindly locking it
In some conditions, we may call pp_hist_setup in a context where
the hist_mutex lock is already taken.

Check if the mutex is already locked prior to re-lock it, otherwise
we may get in a bad situation where we want to acquire it twice,
making the kernel to miserably crash.

Change-Id: Iddae2d7ede1c0a0d45a6c7097af41c4b609f3c30