custom_find7_defconfig: disable CONFIG_PFT as it is unsupported
Disable the drivers/platform/msm/pft.c driver as it is unsupported by
Qualcomm, and opens us up to a wide range of potential attack surfaces
that has not been audited by anyone.

Qualcomm recommends that it be disabled, as it hooks into SELinux in
some "interesting" ways, and the userspace portion of the code is not
even part of the image.

Change-Id: Iaf3884f830b6c807c273c9697e647c5b5f38f973
Ticket: CYNGNOS-3020
Bug: 28588434
Author: Greg Kroah-Hartman <gregkh@google.com>